+++ test/endpoints/users/signup.test.js
@@ -2,7 +2,7 @@
 
 const chai = require('chai');
 const chaiHttp = require('chai-http');
-const app = require('../server');
+const app = require('../../../server');
 
 chai.use(chaiHttp);
 chai.should();
@@ -91,35 +91,6 @@ describe('api/v1/auth/signup', () => {
       chai.request(app).post('/api/v1/auth/signup').send(user).end((err, res) => {
         res.body.should.be.a('object');
         res.body.should.have.status(400);
-        res.body.should.have.property('error');
-        done();
-      });
-    });
-  });
-});
-
-describe('api/v1/auth/signin', () => {
-  describe('POST', () => {
-    it('should sign in a user with correct password', (done) => {
-      const user = {
-        email: 'kelechi22@gmail.com',
-        password: 'hello',
-      };
-      chai.request(app).post('/api/v1/auth/signin').send(user).end((err, res) => {
-        res.should.have.status(200);
-        res.body.should.be.a('object');
-        done();
-      });
-    });
-
-    it('should prevent access to users with incorrect password', (done) => {
-      const user = {
-        email: 'kelechi22@gmail.com',
-        password: 'hel12345',
-      };
-      chai.request(app).post('/api/v1/auth/signin').send(user).end((err, res) => {
-        res.should.have.status(401);
-        res.body.should.be.a('object');
         res.body.should.have.property('error');
         done();
       });
